move errors in eventStore folder
