+++ packages/core/src/eventStore/errors/aggregateNotFound.ts
