+++ packages/core/src/eventStore/errors/eventAlreadyExists.ts
