+++ packages/core/src/eventStore/errors/index.ts
@@ -1,3 +1,3 @@
 export * from './aggregateNotFound';
 export * from './eventAlreadyExists';
-export * from './undefinedStorageAdapterError';
+export * from './undefinedStorageAdapter';
