+++ packages/core/src/eventStore/errors/undefinedStorageAdapter.ts
