+++ packages/core/src/eventStore/eventStore.unit.test.ts
@@ -1,5 +1,4 @@
-import { AggregateNotFoundError } from '~/errors/aggregateNotFound';
-
+import { AggregateNotFoundError } from './errors/aggregateNotFound';
 import {
   CounterAggregate,
   counterCreatedEvent,
