+++ packages/core/src/eventStore/index.ts
@@ -1,3 +1,4 @@
+export * from './errors';
 export * from './eventStore';
 export * from './generics';
 export * from './types';
