+++ packages/core/src/index.ts
@@ -2,13 +2,6 @@ export type { Aggregate } from './aggregate';
 export { EventType } from './event/eventType';
 export type { EventTypeDetail, EventTypesDetails } from './event/eventType';
 export type { EventDetail } from './event/eventDetail';
-export { AggregateNotFoundError } from './errors/aggregateNotFound';
-export type { EventAlreadyExistsError } from './errors/eventAlreadyExists';
-export {
-  isEventAlreadyExistsError,
-  eventAlreadyExistsErrorCode,
-} from './errors/eventAlreadyExists';
-export { UndefinedStorageAdapterError } from './errors/undefinedStorageAdapterError';
 export type { StorageAdapter } from './storageAdapter';
 export type {
   EventsQueryOptions,
@@ -16,8 +9,15 @@ export type {
   ListAggregateIdsOptions,
   ListAggregateIdsOutput,
 } from './storageAdapter';
-export { EventStore } from './eventStore';
+export {
+  AggregateNotFoundError,
+  isEventAlreadyExistsError,
+  eventAlreadyExistsErrorCode,
+  EventStore,
+} from './eventStore';
 export type {
+  EventAlreadyExistsError,
+  UndefinedStorageAdapterError,
   GetAggregateOptions,
   SimulationOptions,
   EventStoreId,
